Move views into a folder
