+++ static/app/views/performance/mobile/appStarts/views/screenSummaryPage.spec.tsx
